+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.2] - 2018-08-25
+### Changed
+- Class`Tokenizer` improved, does recognize delimiters, booleans, integers, real numbers, strings, and identifiers.
+- Spec file `Tokenizer_spec.rb` expanded with more tests.
+
 ## [0.0.1] - 2018-08-25
 ### Added
 - Initial `Tokenizer` class commit